Added files alarm.h alarm.wav main.cpp timer.py
I tried to implement the alarm first in python because I am bad at C/C++
right now. I got the timer to work, but using time.sleep is proving to
be inaccurate. I compared it to time.is and it was not accurate. I need
to find a way to fix this in the future.